+++ srcs/string/ft_itoa.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-char	*ft_itoa(int n)
+char	*ft_itoa(intmax_t n)
 {
 	return (ft_itoa_base(n, 10));
 }
